// src/types/page.ts
export interface BasePageConfig {
    type: 'about' | 'publication' | 'card' | 'text' | 'blog';
    title: string;
    description?: string;
}

export interface PublicationPageConfig extends BasePageConfig {
    type: 'publication';
    source: string;
}

export interface TextPageConfig extends BasePageConfig {
    type: 'text';
    source: string;
}

export interface CardItem {
    title: string;
    subtitle?: string;
    date?: string;
    content?: string;
    tags?: string[];
    link?: string;
    image?: string;
}

export interface CardPageConfig extends BasePageConfig {
    type: 'card';
    items: CardItem[];
}

export interface BlogPost {
    title: string;
    date: string;
    summary: string;
    tags: string[];
}

export interface BlogPageConfig extends BasePageConfig {
    type: 'blog';
    posts: BlogPost[];
}

export interface ProjectItem {
    title: string;
    subtitle?: string;
    date?: string;
    summary?: string;
    content?: string;
    tags?: string[];
    link?: string;
    image?: string;
}

export interface ProjectsPageConfig extends BasePageConfig {
    type: 'projects';
    projects: ProjectItem[];
}